Version 4.7.9
Look for updates using HTTPS.
Options link moved from the page footer to the user menu.
Modal dialogs will resize if the window is resized after they're opened.
Removed old code that determined whether the script was running as a
Firefox addon.
Updated to jQuery 2.2.4.
Workaround for CSS not working on FF55.